backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
chevron==0.14.0
pyjwt==2.10.1
bcrypt>=4.0
supermemory
google-genai
python-dotenv
python-multipart
